feat: add rich text to all required fields
